Parse default macros in custom strings
The game processes two kinds of macros for localization strings: Those that are defined directly on calling the :text function and some default macros (such as for certain key mappings) when a string lookup was successful. BLT does not parse these default macros in custom localization strings and on top of that uses a different format by leaving out the trailing semicolon.

This change adds support for these default macros by looking for macros in the return string in the form of $X; (with the semicolon) and comparing each occurrence with keys in the macro tables. Technically it's more limiting because it doesn't allow single macros to be in strange formats such as $X;$Y; or have any whitespacing in their identifiers, but in any real world scenario this shouldn't be a problem. The only downside to changing the format like this is that any mod using custom localized strings with macros in them would need to be updated, unless the compatibility code is kept in.