Simplify the LocMgr:text function
In my previous commit I wrongfully assumed the default macros are stored including the format characters. This was the case in PD:TH, but has been changed somewhere in the lifetime of PD2. This makes things a lot easier and cleaner. The macro id limitations remain the same.

The deprecation check still makes sense as I still believe the original choice to change the macro format in BLT strings from the vanilla game (leaving out the trailing semicolon) is a bad choice.